requirements.txt
anthropic
google-genai
instructor
ipykernel
jsonref
openai
openpyxl
pandas
pydantic-settings
python-docx
python-dotenv
pytz
pyyaml
streamlit
tabulate
xlrd
xlsxwriter
